requirements.txt
amqp==5.0.9
APScheduler==3.8.1
asgiref==3.5.0
billiard==3.6.4.0
celery==5.2.3
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.0.2
django-apscheduler==0.6.0
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.11.0
django-extensions==3.1.5
django-filter==21.1
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
importlib-metadata==4.11.1
kombu==5.2.3
Markdown==3.3.6
prompt-toolkit==3.0.28
psycopg2==2.9.3
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
schedule==1.1.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
tzlocal==4.1
vine==5.0.0
wcwidth==0.2.5
zipp==3.7.0
